Tea-8838: Trekkjer ut kvar behandlingsrad i ein eigen komponent. Synes det vart lettare å manøvrere når vi skil det sånn
